requirements.txt
llama-index-embeddings-clip
llama-index-core~=0.12.1
pandas~=2.2.3
pillow~=10.4.0
pinecone-client~=5.0.1
pydantic~=2.9.2
SQLAlchemy~=2.0.36
kagglehub
matplotlib
llama_index-llms-gemini
llama_index-experimental
llama_index-embeddings-gemini
llama_index-readers-file
mysql-connector-python~=9.1.0

torch~=2.5.1
protobuf~=4.25.5